removed moved code
